+++ src/components/common/Modal/BottomModal/ModalBase.tsx
@@ -1,5 +1,6 @@
 import { PropsWithChildren } from 'react';
 import classNames from 'classnames/bind';
+import ModalPortal from '@/components/common/Modal/Portal';
 import styles from './BottomModal.module.scss';
 
 export interface ModalProps {
@@ -11,8 +12,10 @@ const cx = classNames.bind(styles);
 
 export default function ModalBase({ className, children }: PropsWithChildren<ModalProps>) {
   return (
-    <div className={cx('modalBase', className)}>
-      <div className={cx('modalContent')}>{children}</div>
-    </div>
+    <ModalPortal>
+      <div className={cx('modalBase', className)}>
+        <div className={cx('modalContent')}>{children}</div>
+      </div>
+    </ModalPortal>
   );
 }
