+++ src/components/common/Modal/BottomModal/index.tsx
@@ -1,6 +1,6 @@
 import classNames from 'classnames/bind';
 import { useModalProps } from '@/hooks/useModal';
-import ModalBase, { ModalProps } from '@/components/common/BottomModal/ModalBase';
+import ModalBase, { ModalProps } from '@/components/common/Modal/BottomModal/ModalBase';
 import styles from './BottomModal.module.scss';
 import { PropsWithChildren } from 'react';
 
@@ -21,7 +21,6 @@ export default function BottomModal({
           handleModalClose();
         }
       }}>
-      <div onClick={handleModalOpen}>모달 열기</div>
       {modalOpen && (
         <ModalBase className={className} onClose={handleModalClose}>
           {children}
@@ -34,20 +33,11 @@ export default function BottomModal({
 {
   /* 사용법
 1. 사용처에서 import useModal from '@/hooks/useModal';
-2. 사용처에서 import ModalPortal from '@/components/Portal';
-3.
-const [isClient, setIsClient] = useState(false);
+2. 사용처에서 import BottomModal from '@/components/common/BottomModal';
+
   const { modalOpen, handleModalOpen, handleModalClose } = useModal();
-
-  useEffect(() => {
-    setIsClient(true);
-  }, []);
-
   return (
     <>
-      {isClient && (
-        <ModalPortal>
-        *모달 사용처 컴포넌트*
             <BottomModal
               className={cx('')}
               modalOpen={modalOpen}
@@ -55,9 +45,6 @@ const [isClient, setIsClient] = useState(false);
               handleModalClose={handleModalClose}>
               *모달 내부 디자인 추가*
             </BottomModal>
-        *모달 사용처 컴포넌트*
-        </ModalPortal>
-      )}
     </>
   );
   사용 예시는 pages > text > my.tsx에서 확인할 수 있습니다.
